backend/requirements.txt
asgiref>=3.7.2
certifi>=2024.6.2
Django>=4.2.23
django-cors-headers>=4.3.1
djangorestframework>=3.15.1
psycopg2-binary>=2.9.9
pytz>=2024.1
sqlparse>=0.5.0
typing_extensions>=4.12.1
wincertstore>=0.2
dj-rest-auth>=5.0.2
django-allauth>=0.63.3
social-auth-app-django>=5.4.0
